requirements.txt
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorclass==2.2.2
coverage==7.11.0
docopt==0.6.2
exceptiongroup==1.3.0
idna==3.11
iniconfig==2.3.0
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pip-upgrader==1.4.15
platformdirs==4.5.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
terminaltables==3.1.10
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
